+++ front_end/src/typescript/World/Helpers/SetOnPlane.test.ts
@@ -2,7 +2,8 @@ import * as THREE from 'three';
 import * as chai from 'chai';
 const expect = chai.expect;
 
-import setOnPlane from '../src/typescript/World/Helpers/SetOnPlane';
+import setOnPlane from './SetOnPlane';
+import {BufferGeometry} from 'three';
 
 describe('setOnPlane Helper', () => {
   describe('Testing planes with normal: y + c = 0', () => {
@@ -23,15 +24,76 @@ describe('setOnPlane Helper', () => {
       const object = new THREE.Object3D();
 
       if (plane.geometry instanceof THREE.BufferGeometry) return;
-      plane.geometry.vertices.forEach(v => {
-        v.y = 5;
-      });
+      for (const vertex of plane.geometry.vertices) vertex.y = 5;
 
       setOnPlane(plane, object, 0, 0);
 
       expect(object.position.x).to.equal(0);
       expect(object.position.y).to.equal(5);
       expect(object.position.z).to.equal(0);
+    });
+
+    it('should return origin + (0, c, 0) when a plane of normal: y + c = 0 is passed with undefined object', () => {
+      const plane = setUpPlane();
+
+      if (plane.geometry instanceof THREE.BufferGeometry) return;
+      for (const vertex of plane.geometry.vertices) vertex.y = 5;
+
+      const y = setOnPlane(plane, undefined, 0, 0);
+
+      expect(y).to.equal(5);
+    });
+
+    describe('object rotation', () => {
+      it('should return the same object position if the plane is flat', () => {
+        const plane = setUpPlane();
+        const object = new THREE.Object3D();
+
+        if (plane.geometry instanceof THREE.BufferGeometry) return;
+        for (const vertex of plane.geometry.vertices) vertex.y = 5;
+
+        const objectRotationAxes: ('x' | 'y' | 'z')[] = ['x', 'y', 'z'];
+        const rotateRelativeToAxes: ('x' | 'y' | 'z')[] = ['x', 'y', 'z'];
+
+        for (const rotationAxis of objectRotationAxes) {
+          for (const rotateRelativeToAxis of rotateRelativeToAxes) {
+            setOnPlane(plane, object, 0, 0, rotationAxis, rotateRelativeToAxis);
+            expect(object.position.x).to.equal(0);
+            expect(object.position.y).to.equal(5);
+            expect(object.position.z).to.equal(0);
+          }
+        }
+      });
+    });
+  });
+
+  describe('Plane has BufferGeometry', () => {
+    it('should throw error', () => {
+      const plane = setUpPlane();
+      const object = new THREE.Object3D();
+
+      if (plane.geometry instanceof THREE.BufferGeometry) return;
+      for (const vertex of plane.geometry.vertices) vertex.y = 5;
+
+      plane.geometry = new BufferGeometry().fromGeometry(plane.geometry);
+
+      expect(setOnPlane.bind(setOnPlane, plane, object, 0, 0)).to.throw(
+        'Buffer planes are not currently supported.'
+      );
+    });
+  });
+
+  describe('x and z are not in plane', () => {
+    it('should throw error', () => {
+      const plane = setUpPlane();
+      const object = new THREE.Object3D();
+
+      if (plane.geometry instanceof THREE.BufferGeometry) return;
+      for (const vertex of plane.geometry.vertices) vertex.y = 5;
+
+      expect(setOnPlane.bind(setOnPlane, plane, object, 10000, 10000)).to.throw(
+        'Position (x,z) is not found in plane.'
+      );
     });
   });
 
@@ -41,9 +103,9 @@ describe('setOnPlane Helper', () => {
       const object = new THREE.Object3D();
 
       if (plane.geometry instanceof THREE.BufferGeometry) return;
-      plane.geometry.vertices.forEach(v => {
-        v.y = Math.random() * 30;
-      });
+      for (const vertex of plane.geometry.vertices) {
+        vertex.y = Math.random() * 30;
+      }
 
       setOnPlane(plane, object, 0, 0);
 
@@ -61,10 +123,10 @@ describe('setOnPlane Helper', () => {
         const object = new THREE.Object3D();
 
         if (plane.geometry instanceof THREE.BufferGeometry) return;
-        plane.geometry.vertices.forEach(v => {
-          if (v.x < 0) v.y = 0;
-          else v.y = 10;
-        });
+        for (const vertex of plane.geometry.vertices) {
+          if (vertex.x < 0) vertex.y = 0;
+          else vertex.y = 10;
+        }
 
         setOnPlane(plane, object, 0, 0);
 
@@ -84,10 +146,10 @@ describe('setOnPlane Helper', () => {
         const maxY = 10;
 
         if (plane.geometry instanceof THREE.BufferGeometry) return;
-        plane.geometry.vertices.forEach(v => {
-          if (v.x < 0) v.y = 0;
-          else v.y = maxY;
-        });
+        for (const vertex of plane.geometry.vertices) {
+          if (vertex.x < 0) vertex.y = 0;
+          else vertex.y = maxY;
+        }
 
         setOnPlane(plane, object, 0, 0);
 
@@ -107,10 +169,10 @@ describe('setOnPlane Helper', () => {
         const maxY = 10;
 
         if (plane.geometry instanceof THREE.BufferGeometry) return;
-        plane.geometry.vertices.forEach(v => {
-          if (v.x < 0) v.y = 0;
-          else v.y = 10;
-        });
+        for (const vertex of plane.geometry.vertices) {
+          if (vertex.x < 0) vertex.y = 0;
+          else vertex.y = 10;
+        }
 
         setOnPlane(plane, object, 2.5, 2.5);